server/pyproject.toml
[tool.poetry]
name = "poker"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
pytest = "^8.0.0"
black = "^24.2.0"
websockets = "^12.0"
tortoise-orm = {extras = ["asyncpg"], version = "^0.20.0"}
transitions = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
